Implement more features in the post page
Added a subscriber list in the footer, which
doesn't do anything yet, but it will allow users to
register themselves for a monthly newsletter.

Split the main post page into different components,
one of them for navigation between articles and another
one for the different sections composiing an article to
give the user information on the current post such as post
date, tags... etc.

Added transition upon loading of an article and a loader that
is a good candidate to become a component that will be usable
in other pages. This allows users to get some visual feedback while
a call is made to apollo.