Add a global footer, extract loading, subscribtion, etc
Added a footer in the global layout that will be used
to display information on reaching out, etc. Standard
web practice.

Extracted the loading spinner into it's own component for
inclusion in other pages that make external requests. This
will improve the user experience.

Extracted the subscriber page into the global component.
Not sure if it will stay there. It is set not to show on
the home page.

Extracted the CSS of the TileContent component because I
wanted to refer to certain SASS variables defined in the
main SASS file. I have heard that it improves performance
to have all the CSS pre-compiled in one file instead of being
spread across multiple components. I think I will start using
seperate style sheets.