Experiment with nuxt generate
Tried to do static generation, end-result is not what
I expected. Some elements render in an odd fashion as their
logic relies on server-fetching timings. As a first, this something
I will want to keep in mind for a future project. It looks like
for static generation it would make more sense to pass the graph
queries payload through the generate key in the nuxt config. At that
point, it looks like a dedicated npm module to abstract that work away
would make sense. Food for thought.