Fix icon loading issue and preadd used icons
Every icon needed to be imported and then defined as a computed property
this was rather verbose which added noise to components and pages. By
adding them in advance as a plugin, they can be referenced without being
imported which makes the components less repetitive. The icon plugin
also preloads the css. Previously, the icons would appear as very large
before being rendered by the browser which affected the user experience.